data penyakit dan presentase sudah muncul program berjalan dengan baik dan benar
